requirements.txt
numpy==1.15.3
Pillow==6.0.0
